--- player/src/windows/mod.rs
@@ -1,5 +1,0 @@
-mod enumeration;
-pub use enumeration::*;
-
-mod run;
-pub use run::*;
